requirements.txt
binance-futures-connector==4.0.0
pandas==2.2.2
ta==0.11.0
python-dotenv==1.0.1
